Create MASH
This is the first working version of MASH, a shell written in C++. In
this version, the shell simply takes in a command and tries to execute
it. There is no error checking, handling, or level. There is also no way
to execute multiple commands, including use of pipes, semicolons, double
ands, or double ors. There are builtin functions for cd, clear, fetch,
and exit. The prompt mimics my default prompt, with the addition of the
prefixed 'M*A*S*H'.