Add Error Handling and Multiple Command Execution
This commit focuses on multiple command execution. To achieve this, the
commands are stored in a structure with associated operator and error
level. The error checking currently detects a single & as a bad
operator, or unknown commands. The error level combined with the
operators ;, &&, and || determine which commands are executed.