Add Pipe Functionality
This commit introduces piping. Currently it only allows for a single
pipe. If two consecutive pipes are entered, an error is thrown. The
piped commands are error checked as normal commands.

I also refactored how the string arguments are converted to char* and
char*[]. I'm not sure how sound it is, but the code is much cleaner.