Add Builtin Function su
This commit introduces the ability to change user. Currently it isn't
very sophisticated, and probably quite insecure. The binary is flagged
as set uid, and owned by root. This allows the program to run 'seteuid'
with the uid of the given username. The effective user id (euid) is
changed, but the real user id (ruid) remains the same as the shell's
caller. I intend to build on this and make it more robust.