Fix Error Handling and Process Exiting
This commit addresses several issues surrounding error handling.
Firstly, I removed the error level for each command, and instead use
return values from the 'executeCommand' function. Secondly, in all areas
of error checking, if an error is found, an error message is thrown and
its associated error code is returned. This value ultimately reaches the
main method and the error level is used in conjuction with the operators
to determine conditional execution. Lastly, the exit statements are
limited to the execution functions, and the exit status of the child
processes are passed to the parent process to be used in error handling.