Fix Order Reversal Inside Parentheses
The first commit of Math Mode had an issue with order of operations
inside parentheses. The issue arose from the assembly of the token
vector inside the parentheses. This was achieved by popping tokens from
the back of the deque, causing the vector to be in reverse order. The
solution was simply to use a reverse algorithm on the vector prior to
performing the calculations on it.