Change User Input to Allow for Arrow Keys
The big change in this commit is how user input is processed. Before it
simply used getline. The problem was that this wouldn't allow for arrow
keys to be used without pressing enter, which prompted an overhaul of
the entire input process. Now it uses a function I stole from
StackOverflow that gets the unbuffered input, prints it to stdout, and
adds it to an input string to be used by the program. This process also
necessitated the need for backspaces to be manually implemented.
Currently the up and down arrow keys delete the user's input, and the
next step will be using them to access the command history.

Other than that, I also removed all uses of stringstreams. I thought
they were necessary, assuming them to act like Java's StringBuilder.
However I discovered that using regular strings is more convenient. I
also removed all calls to the regex library because it slowed
compilation down significantly, and I only used it for whitespace
ignoring, which was easy enough to implement manually.