Implement Aliases from File
This commit introduces the concept of aliases. When the shell starts, or
the user is changed, the user's home folder is checked for a file called
'.mash_aliases'. If found, the contents are parsed and stored in a map.
Any line that doesn't begin with the word 'alias' is ignored, and '#'
can be used as inline comments.
When a command is entered, the command is compared with the aliases. If
a match is found, the command is replaced by the alias. Execution then
continues as normal.

Also, a small change was added to allow 'cd' with no argument to change
to the user's home directory.