Add Alt+. Functionality for Last Argument
This commit is a simple change to how special characters are parsed. In
addition to the arrow keys for history, the key combo Alt+. (or Esc+.)
is now parsed. This combo will insert the last argument from the last
entered command. For example, making and entering a directory can be
done by typing 'mkdir <path>' then simply 'cd <Alt+.>'.